# Enhancement: introduced an actual type system for numeric types (#6)
## Changelog
- Added support for multiple numeric types:
   - `int`/`i32`
   - `uint`/`u32`
   - `float`/`f32`
   - `double`/ `f64`
   - `i8`
   - `i16`
   - `i64`
   - `i128`
   - `u8`
   - `u16`
   - `u64`
   - `u128`

   removed `num`
- Fixed pointers assignment doing a copy assignment instead of assigning the address

Fixes #6

+++ tungsten/src/core/core.cpp
@@ -4,7 +4,7 @@
 // core tungsten functions
 
 extern "C" {
-double shell(char* command) {
+int shell(char* command) {
    if (!command)
       return -1;
